feat: 실제 앱 빌드 및 실행 테스트 추가 (플랫폼별 검증)
- integration_test/app_launch_test.dart 추가
  - 앱 크래시 없이 시작되는지 검증
  - 모든 화면 정상 렌더링 확인
  - 네비게이션 작동 테스트
  - Provider 초기화 검증
  - 화면 회전, 메모리 누수 등 추가 검증

- 플랫폼별 빌드 테스트 스크립트 추가
  - test-ios-build.sh: iOS 시뮬레이터 빌드/실행
  - test-android-build.sh: Android 에뮬레이터 빌드/실행
  - test-web-build.sh: 웹 브라우저 빌드/실행
  - test-all-platforms.sh: 통합 테스트 실행

- CI/CD 파이프라인 통합
  - .github/workflows/build-and-test.yml: PR별 자동 테스트
  - .github/workflows/platform-specific-tests.yml: 일일 플랫폼 테스트

- 테스트 문서 업데이트
  - qa-test-plan.md: 앱 실행 테스트 섹션 추가
  - test-results-98-coverage.md: 플랫폼별 테스트 결과 추가

- 코드 이슈 수정
  - mockDataServiceProvider 중복 선언 해결
  - import 경로 정리

+++ lib/providers/portfolio_provider.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../models/stock_recommendation.dart';
 import '../services/mock_data_service.dart';
+import 'mock_data_provider.dart';
 
 class Position {
   final String id;
@@ -279,6 +280,6 @@ final portfolioStatsProvider = Provider<PortfolioStats>((ref) {
     openPositions: positions.length,
     winningPositions: winningPositions,
     losingPositions: losingPositions,
-    winRate: winRate,
+    winRate: winRate.toDouble(),
   );
 });
